feat: add logic for a button click function
this function is to be called when any of the animal buttons are
clicked.

the main gist is to keep track of a first and second click after which
the tracker is reset. this is achieved by setting an empty variable and
checking whether the variable is empty and if it is this means the
tracker has just been reset or this is the first time the game is
played, in which case we must set it to something to hold the value of
what was just clicked trigering this whole funciton in the first place.

when it does have something in it, we should check whether it matches or
not, if it does match then we reset as this is the 'second time' a
button is clicked, and were replacing the value of the in the button
into an empty string(idk why we have to do this)

when it doesn't match then we just reset the value of the tracker
variable is this is the 'second time'